refactor: Split api.rs
